requirements.txt
Django==2.2.5
djangorestframework==3.9.4
mysqlclient==1.3.13